+++ SourceCode/Kernel/src/interrupt.c
@@ -42,13 +42,6 @@ void disable_interrupt() {
   }
 }
 
-void swi(uint32_t num) {
-  __asm__ __volatile__("push {lr}\n\t"
-                       "mov r0, %0\n\t"
-                       "swi 0x0\n\t"
-                       "pop {pc}\n\t" ::"r"(num));
-}
-
 void __attribute__((interrupt("UNDEF"))) undefined_instruction_handler(void) {}
 
 void __attribute__((interrupt("SWI"))) software_interrupt_handler(uint32_t id) {
